pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdfcraft"
version = "0.1.0"
description = "Python PDF toolkit (CLI + GUI) – your Acrobat-like starter"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name="Your Name", email="you@example.com"}]
license = "MIT"
dependencies = [
    "PyMuPDF>=1.24.9",
    "pikepdf>=9.0.0",
    "PySide6>=6.7.0",
    "typer>=0.12.3",
    "pytesseract>=0.3.10",
    "Pillow>=10.4.0",
    "tqdm>=4.66.4",
    "pyhanko>=0.25.0",
]

[project.scripts]
pdfcraft = "pdfcraft.cli:app"

[tool.setuptools]
packages = ["pdfcraft", "gui"]
